wip on creating documents view with shared sidebar
