+++ react/src/pages/documentsView/DocumentsView.jsx
@@ -1,7 +1,12 @@
 import React from "react";
+import { Box, Button, Container, Typography } from "@mui/material/";
 
 function DocumentsView() {
-  return <div>DocumentsView</div>;
+  return (
+    <>
+      <Typography>Documents</Typography>
+    </>
+  );
 }
 
 export default DocumentsView;
