+++ react/src/pages/queryView/Categories.jsx
